Update manage_site.pp
A lot of people use other software (such as Octopus) to update the folder path when the software version changes.  It would be nice to have a feature that would let us create the virtual host and the initial folder, but if it's been altered by another application (as part of your defined process) that it wouldn't revert the path back.

I'm open to a better way of doing this.