Casing matters
Previously I have processmodel.idletimeoutaction and it was always trying to apply my changes when it saw a difference but it would never actually make the change.  I finally noticed that casing matters and it needs to be processmodel.idleTimeoutAction.  I corrected that as well as also added the check to test to see if the value was NOT undefined.

if $apppool_idle_timeout_action != undef